refactor: move store hooks to its location
