--- src/hooks/index.ts
@@ -1,5 +1,0 @@
-export { useCharacters } from "./use-characters";
-export { useFavorites } from "./use-favorites";
-export { useFilters } from "./use-filters";
-export { useLoading } from "./use-loading";
-export { useSearch } from "./use-search";
