+++ src/store/hooks/index.ts
@@ -1,2 +1,7 @@
-export { useAppState } from "./use-app-state";
 export { useAppDispatch } from "./use-app-dispatch";
+export { useAppState } from "./use-app-state";
+export { useCharacters } from "./use-characters";
+export { useFavorites } from "./use-favorites";
+export { useFilters } from "./use-filters";
+export { useLoading } from "./use-loading";
+export { useSearch } from "./use-search";
